apps/algo-backend/python_scripts/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytrends>=4.9.0
matplotlib>=3.7.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 